+++ .mailmap
@@ -61,3 +61,6 @@ Hank Hua <h.hua23@imperial.ac.uk> hank-hua <146197275+hank-hua@users.noreply.git
 
 #Dom Langridge
 Dom Langridge  <92931431+DomLangridge@users.noreply.github.com> DomLangridge <92931431+DomLangridge@users.noreply.github.com>
+
+#Menai Lamers James
+Menai Lamers James <menai.james@warwick.ac.uk> menailj <77395579+menailj@users.noreply.github.com>
